+++ src/main/java/com/renomad/minum/logging/TestLoggerException.java
@@ -1,0 +1,15 @@
+package com.renomad.minum.logging;
+
+/**
+ * This is an implementation of {@link RuntimeException}, scoped
+ * for the TestLogger.
+ */
+public class TestLoggerException extends RuntimeException {
+
+    /**
+     * See {@link RuntimeException#RuntimeException(String)}
+     */
+    public TestLoggerException(String message) {
+        super(message);
+    }
+}