+++ src/main/java/com/renomad/minum/security/SecurityException.java
@@ -1,0 +1,14 @@
+package com.renomad.minum.security;
+
+/**
+ * Same as {@link RuntimeException} but scoped to the security package
+ */
+public class SecurityException extends RuntimeException {
+
+    /**
+     * See {@link RuntimeException#RuntimeException(String)}
+     */
+    public SecurityException(String message) {
+
+    }
+}