+++ src/main/java/com/renomad/minum/utils/LRUCache.java
@@ -58,5 +58,14 @@ public final class LRUCache<T> extends LinkedHashMap<String, T> {
         return Collections.synchronizedMap(new LRUCache<>(maxSize));
     }
 
+    @Override
+    public boolean equals(Object o) {
+        return super.equals(o);
+    }
+
+    @Override
+    public int hashCode() {
+        return super.hashCode();
+    }
 
 }